+++ user/include/sys/calls.h
@@ -4,7 +4,7 @@
 void kill_me();
 char getchar();
 void putchar(unsigned char c);
-void wait(size_t duration);
-void create_thread(void (*func)(void*), const void* args, size_t args_size);
+void wait(unsigned long duration);
+void create_thread(void (*func)(void*), const void* args, unsigned long args_size);
 
 #endif /* SYSTEM_CALLS_H */
